[AIRFLOW-YYYY] Rename core tests module
